reviews-guide

+++ database/migrations/2026_04_08_076300_create_guide_reviews_table.php
@@ -15,7 +15,7 @@ return new class extends Migration
             $table->id();            
             $table->foreignId('booking_id')->unique()->constrained('guide_bookings')->cascadeOnDelete();
 
-            $table->integer('rating'); // 1–5
+            $table->unsignedTinyInteger('rating'); // 1–5
             $table->text('comment')->nullable();
 
             $table->timestamps();
